Fixed a major issue (+few parsing and undefined errors) of user not remaining logged in after successfully logging in. The problem:
The problem was that when i store a cookie by "document.cookie = name + "=" + value + ";" + expires + ";path=/";" it stores it as "3dpsmoGO0pV51RjfZJJpEMPUJJi/EUuW541aPfh/Um4=" but when i retreve it with "const token = document.cookie.split('; ').find(row => row.startsWith("token=")).split('=')[1];" I get "3dpsmoGO0pV51RjfZJJpEMPUJJi/EUuW541aPfh/Um4" instead.

The cause is apparently something with "base-64 encoding" added the equal sign ('=') as padding for data integrity but decoding/fetching it took the equal sign away.

Now if the user successfully logged in once and go to the login page, it will lead to a message page saying that the user already logged in :D